IntroductionToOOP
In this project, we will introduce the concepts of Object Oriented Programming (OOP) in Java. OOP is a way of modelling programs using objects and classes. In this project, we will demonstrate basic OOP concepts such as class, object, properties and methods by creating a simple Car class. The Car class will contain properties such as the make, model and speed of vehicles and will have behaviours such as accelerating and stopping.